+++ bluesky/qserver.sh
@@ -184,7 +184,7 @@ esac
 # -----------------------------------------------------------------------------
 # :author:    Pete R. Jemian
 # :email:     jemian@anl.gov
-# :copyright: (c) 2017-2022, UChicago Argonne, LLC
+# :copyright: (c) 2017-2024, UChicago Argonne, LLC
 #
 # Distributed under the terms of the Creative Commons Attribution 4.0 International Public License.
 #