+++ CHANGES.rst
@@ -45,7 +45,7 @@ Maintenance
 
 * Drop the (unused now) *stdlogpj* package for configuring Python's *logging*.
 * Environment for bluesky_2024_1.
-
+* Lint the code (for style and syntax errors) using the 'ruff' package.
 
 v1.0.1
 ******