+++ CHANGES.rst
@@ -39,6 +39,8 @@ New Features
   * ophyd-registry
 
 * Build ``registry`` of all ophyd objects.
+* Post a warning if at APS but not on controls subnet.
+
 
 Maintenance
 ------------