+++ CHANGES.rst
@@ -24,12 +24,12 @@ describe future plans.
    1.0.3
    ******
 
-   release expected by 2024-02-01
+   release expected by 2024-04-01
 
 1.0.2
 ******
 
-release expected by 2023-12-31
+release expected by 2024-03-01
 
 New Features
 ------------
@@ -38,9 +38,12 @@ New Features
 
   * ophyd-registry
 
+* Build ``registry`` of all ophyd objects.
+
 Maintenance
 ------------
 
+* Drop the (unused now) *stdlogpj* package for configuring Python's *logging*.
 * Environment for bluesky_2024_1.