+++ CHANGES.rst
@@ -46,7 +46,8 @@ Maintenance
 * Add 'Hello, World!' test to installation checklist.
 * Drop the (unused now) *stdlogpj* package for configuring Python's *logging*.
 * Environment for bluesky_2024_1.
-* Updates (from learned 2-ID) for new installations.
+* Updates (learned from 2-ID) for new installations.
+* Various documentation updates.
 
 
 v1.0.1