+++ CHANGES.rst
@@ -48,8 +48,11 @@ New Features
 Maintenance
 ------------
 
+* Add 'Hello, World!' test to installation checklist.
 * Drop the (unused now) *stdlogpj* package for configuring Python's *logging*.
 * Environment for bluesky_2024_1.
+* Updates (learned from 2-ID) for new installations.
+* Various documentation updates.
 
 
 v1.0.1