+++ CHANGES.rst
@@ -51,9 +51,9 @@ Maintenance
 * Add 'Hello, World!' test to installation checklist.
 * Drop the (unused now) *stdlogpj* package for configuring Python's *logging*.
 * Environment for bluesky_2024_1.
+* Lint the code (for style and syntax errors) using the 'ruff' package.
 * Updates (learned from 2-ID) for new installations.
 * Various documentation updates.
-
 
 v1.0.1
 ******